+++ src/Model/Entity/Post.php
@@ -72,7 +72,7 @@ class Post extends Entity
             return;
         }
 
-        $preview = firstImageFromText($this->_properties['text']);
+        $preview = firstImage($this->_properties['text']);
 
         if ($preview) {
             return $preview;